Engraving: move TabDurationSymbol to separate file
