+++ src/engraving/dom/chordrest.cpp
@@ -52,6 +52,7 @@
 #include "staff.h"
 #include "stafftype.h"
 #include "system.h"
+#include "tabdurationsymbol.h"
 #include "tuplet.h"
 #include "utils.h"
 #include "volta.h"
