+++ src/engraving/dom/factory.cpp
@@ -27,6 +27,7 @@
 #include "accidental.h"
 #include "actionicon.h"
 #include "ambitus.h"
+#include "anchors.h"
 #include "arpeggio.h"
 #include "articulation.h"
 #include "bagpembell.h"
@@ -36,10 +37,10 @@
 #include "box.h"
 #include "bracket.h"
 #include "breath.h"
+#include "capo.h"
 #include "chord.h"
 #include "chordbracket.h"
 #include "chordline.h"
-#include "capo.h"
 #include "deadslapped.h"
 #include "dynamic.h"
 #include "expression.h"
@@ -89,6 +90,7 @@
 #include "score.h"
 #include "segment.h"
 #include "slur.h"
+#include "soundflag.h"
 #include "spacer.h"
 #include "staff.h"
 #include "stafflines.h"
@@ -105,18 +107,16 @@
 #include "systemdivider.h"
 #include "systemlockindicator.h"
 #include "systemtext.h"
-#include "soundflag.h"
+#include "tabdurationsymbol.h"
 #include "tapping.h"
 #include "tempotext.h"
 #include "text.h"
 #include "textline.h"
 #include "tie.h"
 #include "timesig.h"
-#include "anchors.h"
-
-#include "tremolotwochord.h"
-#include "tremolosinglechord.h"
 #include "tremolobar.h"
+#include "tremolosinglechord.h"
+#include "tremolotwochord.h"
 #include "trill.h"
 #include "tripletfeel.h"
 #include "tuplet.h"
