+++ src/engraving/rendering/score/tdraw.cpp
@@ -139,6 +139,7 @@
 #include "dom/systemtext.h"
 #include "dom/soundflag.h"
 
+#include "dom/tabdurationsymbol.h"
 #include "dom/tapping.h"
 #include "dom/tempotext.h"
 #include "dom/text.h"
